+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## main
 
+### Breaking changes
+
+* Adding animation, vertical_align, word_break, display, visibility, & position arguments to the utilities class. `animation: :grow` is now `animation: :hover_grow` this was a change because we changed the class name in primer.
+
+    *Jon Rohan*
+
 ### Misc
 
 * Update contributing guidelines with release instructions.